% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovCoagFullPrediction.R
\docType{package}
\name{CovCoagFullPrediction}
\alias{CovCoagFullPrediction}
\title{CovCoagFullPrediction}
\description{
A package for running prediction model development studies using data in the OMOP CDM
}
